refactor: Split SNS buttons into separate components
- Create separate Twitter and LinkedIn button components

- Add social buttons container component

- Remove old sns-buttons component

- Update HTML files to use new components
