+++ components/registry.js
@@ -2,4 +2,6 @@
 import './header-component.js';
 import './footer-component.js';
 import './dark-mode-toggle.js';
-import './sns-buttons.js'; 
+import './twitter-button.js';
+import './linkedin-button.js';
+import './social-buttons.js';
