+++ index.html
@@ -29,10 +29,10 @@
                     Kuu's Presentations
                 </h1>
                 <div class="flex items-center space-x-4">
-                    <sns-buttons
-                        twitter-url="https://x.com/Fumiya_Kume"
-                        linkedin-url="https://www.linkedin.com/in/kuu-sushi/"
-                    ></sns-buttons>
+                    <social-buttons>
+                        <twitter-button url="https://x.com/Fumiya_Kume"></twitter-button>
+                        <linkedin-button url="https://www.linkedin.com/in/kuu-sushi/"></linkedin-button>
+                    </social-buttons>
                     <dark-mode-toggle></dark-mode-toggle>
                 </div>
             </div>
